requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
mysql-connector-python==8.0.26
python-dotenv==0.19.0
qrcode==7.3
Pillow==8.3.1
gunicorn==20.1.0